% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cetsa_gsea.R
\name{cetsa_gsea}
\alias{cetsa_gsea}
\title{compare_wp}
\usage{
cetsa_gsea(
  hits,
  gene_column = "Genes",
  score_column = "SR",
  condition_column = NULL,
  condition = NULL,
  pos_enrichment = TRUE
)
}
\arguments{
\item{hits}{A data.frame containing the genes id, a score value and preferably a condition column but not necessary.}

\item{gene_column}{The name of the coulumn that contains the genes. Default is 'Genes'.}

\item{score_column}{The name of the coulumn that contains the score values. Default is 'SR'.}

\item{condition_column}{The name of the column that contains the conditions. Default is NULL.}

\item{condition}{The name of the condition you ant to keep. Default is NULL.}

\item{pos_enrichment}{Logical to tell if you want to only look at positive enrichment score.
If FALSE, show only negative.}
}
\value{
A list that contains the results and the plot.
}
\description{
Function to run enrichment analysis on your hits from the wiki-pathway database and return
a plot that compare the pathways found between your conditions.
}
\details{
For now, only Homo Sapiens is supported.
}
\seealso{
\code{\link{clusterProfiler}}
}
