% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chemarrest}
\alias{chemarrest}
\title{Chemarrest fold-change dataset}
\format{
## `chemarrest`
A data frame with 2777 rows and 59 columns:
\describe{
  \item{id}{Uniprot protein ID}
  \item{description}{Protein description}
  \item{37C_B1_G1S}{Log2 fold-change G1S vs G1S phase at 37°C in the first bioreplicate}
  \item{37C_B1_PM}{Log2 fold-change PM vs G1S phase at 37°C in the first bioreplicate}
  \item{37C_B1_S}{Log2 fold-change S vs G1S phase at 37°C in the first bioreplicate}
  \item{37C_B2_G1S}{Log2 fold-change G1S vs G1S phase at 37°C in the second bioreplicate}
  \item{37C_B2_PM}{Log2 fold-change PM vs G1S phase at 37°C in the second bioreplicate}
  \item{37C_B2_S}{Log2 fold-change S vs G1S phase at 37°C in the second bioreplicate}
  \item{37C_B3_G1S}{Log2 fold-change G1S vs G1S phase at 37°C in the third bioreplicate}
  \item{37C_B3_PM}{Log2 fold-change PM vs G1S phase at 37°C in the third bioreplicate}
  \item{37C_B3_S}{Log2 fold-change S vs G1S phase at 37°C in the third bioreplicate}
  \item{47C_B1_G1S}{Log2 fold-change G1S vs G1S phase at 47°C in the first bioreplicate}
  \item{47C_B1_PM}{Log2 fold-change PM vs G1S phase at 47°C in the first bioreplicate}
  \item{47C_B1_S}{Log2 fold-change S vs G1S phase at 47°C in the first bioreplicate}
  \item{47C_B2_G1S}{Log2 fold-change G1S vs G1S phase at 47°C in the second bioreplicate}
  \item{47C_B2_PM}{Log2 fold-change PM vs G1S phase at 47°C in the second bioreplicate}
  \item{47C_B2_S}{Log2 fold-change S vs G1S phase at 47°C in the second bioreplicate}
  \item{47C_B3_G1S}{Log2 fold-change G1S vs G1S phase at 47°C in the third bioreplicate}
  \item{47C_B3_PM}{Log2 fold-change PM vs G1S phase at 47°C in the third bioreplicate}
  \item{47C_B3_S}{Log2 fold-change S vs G1S phase at 47°C in the third bioreplicate}
  \item{50C_B1_G1S}{Log2 fold-change G1S vs G1S phase at 50°C in the first bioreplicate}
  \item{50C_B1_PM}{Log2 fold-change PM vs G1S phase at 50°C in the first bioreplicate}
  \item{50C_B1_S}{Log2 fold-change S vs G1S phase at 50°C in the first bioreplicate}
  \item{50C_B2_G1S}{Log2 fold-change G1S vs G1S phase at 50°C in the second bioreplicate}
  \item{50C_B2_PM}{Log2 fold-change PM vs G1S phase at 50°C in the second bioreplicate}
  \item{50C_B2_S}{Log2 fold-change S vs G1S phase at 50°C in the second bioreplicate}
  \item{50C_B3_G1S}{Log2 fold-change G1S vs G1S phase at 50°C in the third bioreplicate}
  \item{50C_B3_PM}{Log2 fold-change PM vs G1S phase at 50°C in the third bioreplicate}
  \item{50C_B3_S}{Log2 fold-change S vs G1S phase at 50°C in the third bioreplicate}
  \item{52C_B1_G1S}{Log2 fold-change G1S vs G1S phase at 52°C in the first bioreplicate}
  \item{52C_B1_PM}{Log2 fold-change PM vs G1S phase at 52°C in the first bioreplicate}
  \item{52C_B1_S}{Log2 fold-change S vs G1S phase at 52°C in the first bioreplicate}
  \item{52C_B2_G1S}{Log2 fold-change G1S vs G1S phase at 52°C in the second bioreplicate}
  \item{52C_B2_PM}{Log2 fold-change PM vs G1S phase at 52°C in the second bioreplicate}
  \item{52C_B2_S}{Log2 fold-change S vs G1S phase at 52°C in the second bioreplicate}
  \item{52C_B3_G1S}{Log2 fold-change G1S vs G1S phase at 52°C in the third bioreplicate}
  \item{52C_B3_PM}{Log2 fold-change PM vs G1S phase at 52°C in the third bioreplicate}
  \item{52C_B3_S}{Log2 fold-change S vs G1S phase at 52°C in the third bioreplicate}
  \item{54C_B1_G1S}{Log2 fold-change G1S vs G1S phase at 54°C in the first bioreplicate}
  \item{54C_B1_PM}{Log2 fold-change PM vs G1S phase at 54°C in the first bioreplicate}
  \item{54C_B1_S}{Log2 fold-change S vs G1S phase at 54°C in the first bioreplicate}
  \item{54C_B2_G1S}{Log2 fold-change G1S vs G1S phase at 54°C in the second bioreplicate}
  \item{54C_B2_PM}{Log2 fold-change PM vs G1S phase at 54°C in the second bioreplicate}
  \item{54C_B2_S}{Log2 fold-change S vs G1S phase at 54°C in the second bioreplicate}
  \item{54C_B3_G1S}{Log2 fold-change G1S vs G1S phase at 54°C in the third bioreplicate}
  \item{54C_B3_PM}{Log2 fold-change PM vs G1S phase at 54°C in the third bioreplicate}
  \item{54C_B3_S}{Log2 fold-change S vs G1S phase at 54°C in the third bioreplicate}
  \item{57C_B1_G1S}{Log2 fold-change G1S vs G1S phase at 57°C in the first bioreplicate}
  \item{57C_B1_PM}{Log2 fold-change PM vs G1S phase at 57°C in the first bioreplicate}
  \item{57C_B1_S}{Log2 fold-change S vs G1S phase at 57°C in the first bioreplicate}
  \item{57C_B2_G1S}{Log2 fold-change G1S vs G1S phase at 57°C in the second bioreplicate}
  \item{57C_B2_PM}{Log2 fold-change PM vs G1S phase at 57°C in the second bioreplicate}
  \item{57C_B2_S}{Log2 fold-change S vs G1S phase at 57°C in the second bioreplicate}
  \item{57C_B3_G1S}{Log2 fold-change G1S vs G1S phase at 57°C in the third bioreplicate}
  \item{57C_B3_PM}{Log2 fold-change PM vs G1S phase at 57°C in the third bioreplicate}
  \item{57C_B3_S}{Log2 fold-change S vs G1S phase at 57°C in the third bioreplicate}
  \item{sumUniPeps}{Median of the number of unique peptides from all temperatures}
  \item{sumPSMs}{Median of the PSMs from all temperatures}
  \item{countNum}{Median of the abundance count from all reference channel}
}
}
\source{
This data were obtained from 7.	Dai, L., T. Zhao, X. Bisteau et al. (2018)
        Modulation of Protein-Interaction States through the Cell Cycle.
        Cell, 173, 1481-1494.e13.
}
\usage{
chemarrest
}
\description{
An IMPRINTS-CETSA dataset containing the log2 fold-change between each bio replicate from
different cell cycle phase.
}
\keyword{datasets}
