% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_prot_loop.R
\name{com_protein_loop}
\alias{com_protein_loop}
\title{com_protein_loop}
\usage{
com_protein_loop(data)
}
\arguments{
\item{data}{A named list containing at least two elements which contains character of numeric elements}
}
\value{
A list containing the common and unique element of the whole list
}
\description{
Function to know if each element of each element of a list is unique or duplicated; and if so, how many times.
}
\examples{
# with numeric
x <- list("a" = c(1,2,3,5), "b" = c(55,1,8,99,65,3,4), "c" = c(7,0,1), "d" = c(55,1,65,3,5))
com_protein_loop(x)

# with character, here these are proteins
x <- mineCETSAapp::drug_data
com_protein_loop(lapply(x$data, function(x) x$id))

}
