% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_enrich.R
\name{compare_enrich}
\alias{compare_enrich}
\title{compare_enrich}
\usage{
compare_enrich(
  hits,
  gene_column = "Gene",
  treatment_column = NULL,
  species = c("human", "mouse"),
  n_pathway = 5,
  pval_cutoff = 0.01,
  minGSSize = 3,
  database = c("WikiPathway", "KEGG", "GO", "CETSA")
)
}
\arguments{
\item{hits}{The hitlist; a data.frame containing the genes id and preferably a treatment column but not necessary.}

\item{gene_column}{The name of the coulumn that contains the genes. Default is 'Gene'.}

\item{treatment_column}{The name of the column that contains the treatments. Default is NULL.}

\item{species}{Specify the species. Currently, only 'human' and 'mouse' are available.}

\item{n_pathway}{Number of pathway to show on plot. Default is 5.
For more info you, see \code{\link{compareCluster}}.}

\item{pval_cutoff}{The p-value cutoff for the enrichment analysis.}

\item{minGSSize}{minimal size of each gene set for analyzing. Default here is 3}

\item{database}{Specify the database. Currently, WikiPathway, KEGG, GO and CETSA are available.}
}
\value{
A list that contains the results and the plot.
}
\description{
Function to run over representation analysis on your hits and return
a plot that compare the pathways found between your treatments.
}
\seealso{
\code{\link{clusterProfiler}}
}
