% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_wp.R
\name{compare_wp}
\alias{compare_wp}
\title{compare_wp}
\usage{
compare_wp(hits, gene_column = "Genes", condition_column = NULL, n_pathway = 5)
}
\arguments{
\item{hits}{The hitlist; a data.frame containing the genes id and preferably a condition column but not necessary.}

\item{gene_column}{The name of the coulumn that contains the genes. Default is 'Genes'.}

\item{condition_column}{The name of the column that contains the conditions. Default is NULL.}

\item{n_pathway}{Number of pathway to show on plot. Default is 5.
For more info you, see \code{\link{compareCluster}}.}
}
\value{
A list that contains the results and the plot.
}
\description{
Function to run enrichment analysis on your hits from the wiki-pathway database and return
a plot that compare the pathways found between your conditions.
}
\details{
For now, only Homo Sapiens is supported.
}
\seealso{
\code{\link{clusterProfiler}}
}
