% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elu_example1_raw_joined}
\alias{elu_example1_raw_joined}
\title{Raw Protein.Group from the shortened  elutriation dataset}
\format{
## `elu_example1_raw_joined`
A data frame with 3170 rows and 16 columns:
\describe{
  \item{id}{Uniprot protein ID}
  \item{description}{Protein description}
  \item{condition}{}
  \item{B1_G1}{Raw protein abundance of the first bioreplicate of the G1 phase}
  \item{B1_S}{Raw protein abundance of the first bioreplicate of the S phase}
  \item{B1_G2}{Raw protein abundance of the first bioreplicate of the G2 phase}
  \item{B2_G1}{Raw protein abundance of the second bioreplicate of the G1 phase}
  \item{B2_S}{Raw protein abundance of the second bioreplicate of the S phase}
  \item{B2_G2}{Raw protein abundance of the second bioreplicate of the G2 phase}
  \item{B3_S}{Raw protein abundance of the third bioreplicate of the S phase}
  \item{B3_G1}{Raw protein abundance of the third bioreplicate of the G1 phase}
  \item{B3_G2}{Raw protein abundance of the third bioreplicate of the G2 phase}
  \item{Mix}{Raw protein abundance of the Mix channel}
  \item{sumUniPeps}{Number of unique peptides}
  \item{sumPSMs}{PSMs}
  \item{countNum}{Abundance count from reference channel}
}
}
\source{
This data were obtained from 7.	Dai, L., T. Zhao, X. Bisteau et al. (2018)
        Modulation of Protein-Interaction States through the Cell Cycle.
        Cell, 173, 1481-1494.e13.
}
\usage{
elu_example1_raw_joined
}
\description{
A data frame containing the six Protein.Group files from the raw data of the elutriation dataset.
These raw data have been processed with Proteome Discovers version 2.4 and shortened  to
approximately 500 proteins to showcase the functionnalities of the app and show the expected format.
}
\keyword{datasets}
