% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elu_example2_raw_cleaned}
\alias{elu_example2_raw_cleaned}
\title{Filtered Protein.Group from the shortened  elutriation dataset}
\format{
## `elu_example2_raw_cleaned`
A data frame with 3167 rows and 15 columns:
\describe{
  \item{id}{Uniprot protein ID}
  \item{description}{Protein description}
  \item{condition}{}
  \item{B1_G1}{Raw protein abundance of the first bioreplicate of the G1 phase}
  \item{B1_S}{Raw protein abundance of the first bioreplicate of the S phase}
  \item{B1_G2}{Raw protein abundance of the first bioreplicate of the G2 phase}
  \item{B2_G1}{Raw protein abundance of the second bioreplicate of the G1 phase}
  \item{B2_S}{Raw protein abundance of the second bioreplicate of the S phase}
  \item{B2_G2}{Raw protein abundance of the second bioreplicate of the G2 phase}
  \item{B3_S}{Raw protein abundance of the third bioreplicate of the S phase}
  \item{B3_G1}{Raw protein abundance of the third bioreplicate of the G1 phase}
  \item{B3_G2}{Raw protein abundance of the third bioreplicate of the G2 phase}
  \item{sumUniPeps}{Number of unique peptides}
  \item{sumPSMs}{PSMs}
  \item{countNum}{Abundance count from reference channel}
}
}
\source{
This data were obtained from 7.	Dai, L., T. Zhao, X. Bisteau et al. (2018)
        Modulation of Protein-Interaction States through the Cell Cycle.
        Cell, 173, 1481-1494.e13.
}
\usage{
elu_example2_raw_cleaned
}
\description{
A data frame containing the cleaned and filtered raw protein abundance using the \code{ms_clean} function
from \code{IMPRINTS.CETSA} package.
}
\details{
See \code{\link{elu_example1_raw_joined}}
}
\keyword{datasets}
