% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elu_example4_pre_norm}
\alias{elu_example4_pre_norm}
\title{Rearranged protein abundance from the shortened elutriation dataset}
\format{
## `elu_example4_pre_norm`
A data frame with 521 rows and 59 columns:
\describe{
  \item{id}{Uniprot protein ID}
  \item{description}{Protein description}
  \item{37C_B1_G1}{Raw protein abundance in G1 phase at 37°C in the first bioreplicate}
  \item{37C_B1_G2}{Raw protein abundance in G2 phase at 37°C in the first bioreplicate}
  \item{37C_B1_S}{Raw protein abundance in S phase at 37°C in the first bioreplicate}
  \item{37C_B2_G1}{Raw protein abundance in G1 phase at 37°C in the second bioreplicate}
  \item{37C_B2_G2}{Raw protein abundance in G2 phase at 37°C in the second bioreplicate}
  \item{37C_B2_S}{Raw protein abundance in S phase at 37°C in the second bioreplicate}
  \item{37C_B3_G1}{Raw protein abundance in G1 phase at 37°C in the third bioreplicate}
  \item{37C_B3_G2}{Raw protein abundance in G2 phase at 37°C in the third bioreplicate}
  \item{37C_B3_S}{Raw protein abundance in S phase at 37°C in the third bioreplicate}
  \item{47C_B1_G1}{Raw protein abundance in G1 phase at 47°C in the first bioreplicate}
  \item{47C_B1_G2}{Raw protein abundance in G2 phase at 47°C in the first bioreplicate}
  \item{47C_B1_S}{Raw protein abundance in S phase at 47°C in the first bioreplicate}
  \item{47C_B2_G1}{Raw protein abundance in G1 phase at 47°C in the second bioreplicate}
  \item{47C_B2_G2}{Raw protein abundance in G2 phase at 47°C in the second bioreplicate}
  \item{47C_B2_S}{Raw protein abundance in S phase at 47°C in the second bioreplicate}
  \item{47C_B3_G1}{Raw protein abundance in G1 phase at 47°C in the third bioreplicate}
  \item{47C_B3_G2}{Raw protein abundance in G2 phase at 47°C in the third bioreplicate}
  \item{47C_B3_S}{Raw protein abundance in S phase at 47°C in the third bioreplicate}
  \item{50C_B1_G1}{Raw protein abundance in G1 phase at 50°C in the first bioreplicate}
  \item{50C_B1_G2}{Raw protein abundance in G2 phase at 50°C in the first bioreplicate}
  \item{50C_B1_S}{Raw protein abundance in S phase at 50°C in the first bioreplicate}
  \item{50C_B2_G1}{Raw protein abundance in G1 phase at 50°C in the second bioreplicate}
  \item{50C_B2_G2}{Raw protein abundance in G2 phase at 50°C in the second bioreplicate}
  \item{50C_B2_S}{Raw protein abundance in S phase at 50°C in the second bioreplicate}
  \item{50C_B3_G1}{Raw protein abundance in G1 phase at 50°C in the third bioreplicate}
  \item{50C_B3_G2}{Raw protein abundance in G2 phase at 50°C in the third bioreplicate}
  \item{50C_B3_S}{Raw protein abundance in S phase at 50°C in the third bioreplicate}
  \item{52C_B1_G1}{Raw protein abundance in G1 phase at 52°C in the first bioreplicate}
  \item{52C_B1_G2}{Raw protein abundance in G2 phase at 52°C in the first bioreplicate}
  \item{52C_B1_S}{Raw protein abundance in S phase at 52°C in the first bioreplicate}
  \item{52C_B2_G1}{Raw protein abundance in G1 phase at 52°C in the second bioreplicate}
  \item{52C_B2_G2}{Raw protein abundance in G2 phase at 52°C in the second bioreplicate}
  \item{52C_B2_S}{Raw protein abundance in S phase at 52°C in the second bioreplicate}
  \item{52C_B3_G1}{Raw protein abundance in G1 phase at 52°C in the third bioreplicate}
  \item{52C_B3_G2}{Raw protein abundance in G2 phase at 52°C in the third bioreplicate}
  \item{52C_B3_S}{Raw protein abundance in S phase at 52°C in the third bioreplicate}
  \item{54C_B1_G1}{Raw protein abundance in G1 phase at 54°C in the first bioreplicate}
  \item{54C_B1_G2}{Raw protein abundance in G2 phase at 54°C in the first bioreplicate}
  \item{54C_B1_S}{Raw protein abundance in S phase at 54°C in the first bioreplicate}
  \item{54C_B2_G1}{Raw protein abundance in G1 phase at 54°C in the second bioreplicate}
  \item{54C_B2_G2}{Raw protein abundance in G2 phase at 54°C in the second bioreplicate}
  \item{54C_B2_S}{Raw protein abundance in S phase at 54°C in the second bioreplicate}
  \item{54C_B3_G1}{Raw protein abundance in G1 phase at 54°C in the third bioreplicate}
  \item{54C_B3_G2}{Raw protein abundance in G2 phase at 54°C in the third bioreplicate}
  \item{54C_B3_S}{Raw protein abundance in S phase at 54°C in the third bioreplicate}
  \item{57C_B1_G1}{Raw protein abundance in G1 phase at 57°C in the first bioreplicate}
  \item{57C_B1_G2}{Raw protein abundance in G2 phase at 57°C in the first bioreplicate}
  \item{57C_B1_S}{Raw protein abundance in S phase at 57°C in the first bioreplicate}
  \item{57C_B2_G1}{Raw protein abundance in G1 phase at 57°C in the second bioreplicate}
  \item{57C_B2_G2}{Raw protein abundance in G2 phase at 57°C in the second bioreplicate}
  \item{57C_B2_S}{Raw protein abundance in S phase at 57°C in the second bioreplicate}
  \item{57C_B3_G1}{Raw protein abundance in G1 phase at 57°C in the third bioreplicate}
  \item{57C_B3_G2}{Raw protein abundance in G2 phase at 57°C in the third bioreplicate}
  \item{57C_B3_S}{Raw protein abundance in S phase at 57°C in the third bioreplicate}
  \item{sumUniPeps}{Median of the number of unique peptides from all temperatures}
  \item{sumPSMs}{Median of the PSMs from all temperatures}
  \item{countNum}{Median of the abundance count from all reference channel}
}
}
\source{
This data were obtained from 7.	Dai, L., T. Zhao, X. Bisteau et al. (2018)
        Modulation of Protein-Interaction States through the Cell Cycle.
        Cell, 173, 1481-1494.e13.
}
\usage{
elu_example4_pre_norm
}
\description{
A data frame containing the raw abundance protein from the shortened elutriation dataset put in the right
IMPRINTS-CETSA format.
}
\details{
See \code{\link{elu_example1_raw_joined}}
}
\keyword{datasets}
