% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed_search.R
\name{find_in_pubmed}
\alias{find_in_pubmed}
\title{find_in_pubmed}
\usage{
find_in_pubmed(
  data,
  feat = "PI3K",
  imp_by_hitlist = FALSE,
  treatment = "",
  language = NULL,
  year_rg = NULL,
  your_API = NULL,
  save_word = TRUE,
  newfolder_name = "PI3K_pubmed_search"
)
}
\arguments{
\item{data}{The word you want to search a publication on. It can be a file path, a character vector, a list or
the dataframe hitlist output from the hitlist function.}

\item{feat}{A second word that will be match to every word from data, can be NULL.}

\item{imp_by_hitlist}{Logical to tell if the data are in the hitlist format.}

\item{treatment}{A character vector for selecting specific treatment. Use when data are in hitlist format.}

\item{language}{A character to tell if you want to select publications in a specific language like 'english'.}

\item{year_rg}{A character to tell a range of year of publication. The format is '2000:2023'}

\item{your_API}{A character that is your NCBI API key. If you don't have an account, use NULL.}

\item{save_word}{A logical to tell if you want to save each publication found (title and abstract)
in a word file for each query.}

\item{newfolder_name}{The name of the folder that will be created for saving the word documents}
}
\value{
Dataframe which contains the proteins and if they give a result or not
}
\description{
Function to check if there is already some publications in Pubmed on your subject.
If so, export word files of the publications
}
\examples{
library(IMPRINTS.CETSA.app)
find_in_pubmed("Lamin B1", imp_by_hitlist = FALSE)

}
\seealso{
\code{\link{hitlist}} and \code{\link{pmApi2df}} for more details
}
