% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_concept_net.R
\name{gene_concept_net}
\alias{gene_concept_net}
\title{gene_concept_net}
\usage{
gene_concept_net(
  hits,
  gene_column = "Gene",
  score_column = "IS",
  treatment_column = NULL,
  treatment = NULL,
  species = c("human", "mouse"),
  pval_cutoff = 0.01,
  minGSSize = 3,
  database = c("WikiPathway", "KEGG", "GO", "CETSA")
)
}
\arguments{
\item{hits}{A data.frame containing the genes id, a score value and preferably a treatment column but not necessary.}

\item{gene_column}{The name of the column that contains the genes. Default is 'Gene'.}

\item{score_column}{The name of the column that contains the score values. Default is 'IS'.}

\item{treatment_column}{The name of the column that contains the treatments. Default is NULL.}

\item{treatment}{The name of the treatment you ant to keep. Default is NULL.}

\item{species}{Specify the species. Currently, only 'human' and 'mouse' are available.}

\item{pval_cutoff}{The p-value cutoff for the gene concept network.}

\item{minGSSize}{minimal size of each gene set for analyzing. default here is 3}

\item{database}{Specify the database. Currently, WikiPathway, KEGG, GO and CETSA are available.}
}
\value{
The gene concept network plot.
}
\description{
Function to run over representation analysis on your hits and return
a network plot.
}
\seealso{
\code{\link{clusterProfiler}}
}
