% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_GO.R
\name{Get_GO}
\alias{Get_GO}
\title{Get_GO}
\usage{
Get_GO(anno, enrich = FALSE, all_cat = TRUE, sing_cat = "Component")
}
\arguments{
\item{anno}{The output data from  get_annotations or get_enrichment functions from STRINGdb package.}

\item{enrich}{A logical to tell whether your data is the output from
get_enrichment function from STRINGdb package or not}

\item{all_cat}{A logical to tell if you want the results from all the database or not}

\item{sing_cat}{A character; if all_cat is set to FALSE, will select only this database}
}
\value{
A list of the results
}
\description{
Function to get the result list from the get_annotations or get_enrichment functions from STRINGdb
package.
}
\seealso{
\code{\link{STRINGdb}}
}
