% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_treat_level.R
\name{get_treat_level}
\alias{get_treat_level}
\title{get_treat_level}
\usage{
get_treat_level(data)
}
\arguments{
\item{data}{An IMPRINTS-CETSA dataset; after imprints_normalization or imprints_caldiff for example.}
}
\value{
A character vector which contains all the treatment level from the data
}
\description{
Function to get the treatment level of the CETSA data.
}
