% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hit_for_cell.R
\name{hit_for_cell}
\alias{hit_for_cell}
\title{hit_for_cell}
\usage{
hit_for_cell(HIT, organism = c("HUMAN", "MOUSE"), cell_coordinate = TRUE)
}
\arguments{
\item{HIT}{Usually, the output from hitlist function; but just has to be a data frame with the
columns id, treatment and category}

\item{organism}{The organism on which you performed your experiment. For now, only 'HUMAN' and 'MOUSE' are available.}

\item{cell_coordinate}{Logical to randomly assign or not coordinate on the cell png corresponding to
the mapped cellular location. Default is TRUE.}
}
\value{
Dataframe (format for hit_plotcell)
}
\description{
Function to get subcellular location based on the database from protein Atlas, and coordinates
of each protein on the plot cell (img2) according to their location
}
\seealso{
\code{\link{hitplot_cell}}
}
