% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hitlist_chemarrest}
\alias{hitlist_chemarrest}
\title{Hitlist from the chemarrest dataset}
\format{
## `hitlist_chemarrest`
A data frame with 732 rows and 3 columns:
\describe{
  \item{id}{Uniprot protein ID}
  \item{treatment}{The treatment in which it was identified as a hit}
  \item{category}{Category of the hit}
}
}
\source{
This data were obtained from 7.	Dai, L., T. Zhao, X. Bisteau et al. (2018)
        Modulation of Protein-Interaction States through the Cell Cycle.
        Cell, 173, 1481-1494.e13.
}
\usage{
hitlist_chemarrest
}
\description{
An hitlist from the chemarrest dataset obtained with the \code{imprints_score} function
from the package \code{IMPRINTS.CETSA} using default parameters except useMAD set to TRUE.
}
\keyword{datasets}
