% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitlist_stat.R
\name{hitlist_stat}
\alias{hitlist_stat}
\title{hitlist_stat}
\usage{
hitlist_stat(
  data = NULL,
  control = NULL,
  rem_proteomic = NULL,
  basetemp = "37C",
  RSS_cutoff = 0.985,
  ErrScore = NULL,
  use_prompt = TRUE,
  exported = FALSE,
  format = c("xlsx", "txt", "csv")
)
}
\arguments{
\item{data}{The input data set from which categorization is performed on and hitlist is produced from, ie. caldiff output.
You can also type a file path. The file must be in txt, xlsx or csv file. If in csv, check
if in the description column there are no commas. Otherwise, it could lead to a misreading of the data.
(The function will remind you to do so)}

\item{control}{The name(s) of the control(s). If several, separate them by '|'.}

\item{rem_proteomic}{The name of the proteomic channel (example : '36C'). Can be NULL.}

\item{basetemp}{The name of the base temperature (example : '37C').}

\item{RSS_cutoff}{The residual sum of squares cutoff between the base temperature and the others.
This cutoff is for differentitate the CN and the CC.
The RSS is reduced to a score between 0 and 1 (1/(1+RSS)).
The more it's close to one, the more the RSS is close to 0. So the more you choose
a cutoff close to one, the more you will consider small variations between the base temperatures and the
others as thermal stability change (categorized as CC then). Default is 0.985.}

\item{ErrScore}{The error score threshold, a score between 0 and 1; it is calculated thanks to the  standard error.
From the SE, you will get the error bar of each mean value for each protein for each temperature and 
conditions. Then, from this error value, the score is calculated as follow : 
\code{1/(1+abs((error/mean)))}. So the more it's close to one, the you're confident about your measurment.
You can choose a value or not. If not, the algorithm will plot the histogramm of the error  score of the hits
and take the error score that gives the highest frequency. If several conditions, will take the
maximum threshold value found between the conditions.}

\item{use_prompt}{Logical to tell if you want to use the prompt function (added for the shiny app)}

\item{exported}{Logical to tell if you want to export the results. Use it when use_prompt is FALSE.}

\item{format}{The format of the saved files. Format available : xlsx, txt or csv.}
}
\value{
A list which contains the hitlist, the summary and the NN
}
\description{
Function to categorize protein according to their Expression/Stability change
based on two treatment conditions and characteristic conditions for fold change.
}
\examples{
\dontrun{
library(mineCETSAapp)
PI3K_data <- mineCETSAapp::PI3K1h6h_file
HITS <- hitlist_stat(PI3K_data, control = "Vehicle1h|Vehicle6h",
                     rem_proteomic = "36C", format = "xlsx")
}

}
