% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMPRINTS.CETSA.app.R
\docType{package}
\name{IMPRINTS.CETSA.app}
\alias{IMPRINTS.CETSA.app-package}
\alias{IMPRINTS.CETSA.app}
\title{IMPRINTS.CETSA.app Package}
\description{
Contains a shiny app for IMPRINTS.CETSA R package and
several functions to analyze IMPRINTS CETSA data in a user friendly way.
}
\author{
Marc-Antoine GERAULT \email{marco.gerault@gmail.com}
}
