% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprints_average_sh.R
\name{imprints_average_sh}
\alias{imprints_average_sh}
\title{imprints_average_sh}
\usage{
imprints_average_sh(data, savefile = TRUE)
}
\arguments{
\item{data}{dataset after calculating the relative protein abundance differences}

\item{savefile}{A logical to tell if you want save the results or not}
}
\value{
A dataframe
}
\description{
Function to calculate the averaged signals from the IMPRTINTS-CETSA result
It is totally based on the function ms_2D_average from the mineCETSA package.
}
\seealso{
\code{\link{ms_2D_caldiff}}
}
