% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprints_barplotting_simprof.R
\name{imprints_barplotting_simprof}
\alias{imprints_barplotting_simprof}
\title{imprints_barplotting_simprof}
\usage{
imprints_barplotting_simprof(
  data,
  data_average = NULL,
  treatmentlevel = "Buparlisib6h",
  protein_profile = "P85037",
  setlevel = NULL,
  corrtable = NULL,
  plotseq = NULL,
  printBothName = TRUE,
  printGeneName = FALSE,
  score_threshold = 0.9,
  max_na_prow = 0,
  use_score = "euclidean",
  pfdatabase = FALSE,
  witherrorbar = TRUE,
  layout = NULL,
  colorpanel = "#18FF00",
  usegradient = FALSE,
  colorgradient = c("#4575B4", "ivory", "#D73027"),
  linegraph = FALSE,
  log2scale = TRUE,
  ratio = 0.6,
  ret_plot = FALSE,
  withpopup = FALSE,
  continue = TRUE,
  modvar = "",
  got_it = FALSE,
  save_prlist = TRUE,
  save_pdf = TRUE,
  toplabel = "IMPRINTS-CETSA bar plotting",
  leftlabel = "",
  bottomlabel = "",
  pdfname = "barplot",
  pdfheight = 12,
  pdfwidth = 12
)
}
\arguments{
\item{data}{dataset after imprints_caldiff to plot.}

\item{data_average}{dataset after imprints_average. If null, will get it from data.}

\item{treatmentlevel}{A single character element which corresponds to one of the condition from the data}

\item{protein_profile}{A single character element which corresponds to one
of the protein from the data that you want the similar profile}

\item{setlevel}{a vector of set information if any, such as c("M13","M16")}

\item{corrtable}{A correlation table}

\item{plotseq}{a vector of plots arragement sequence (in composite ID)}

\item{printBothName}{A logical to tell if you want to print the both protein names on the plot}

\item{printGeneName}{A logical to tell if you want to print the gene names on the plot}

\item{score_threshold}{A numeric value to indicate the threshold, default set to 0.9}

\item{max_na_prow}{An integer indicating the maximun number of missing value for one protein}

\item{use_score}{A single character element that define the method score. Method available : 'euclidean' or 'pearson'}

\item{pfdatabase}{A logical for using pdf database or not}

\item{witherrorbar}{A logical to print or not the error bar on the plot}

\item{layout}{a vector indicating the panel layout for multi-panel plots per page,
default value is c(2,3) for set containing data, otherwise c(4,3), use when save_pdf = TRUE}

\item{colorpanel}{a vector of customizable color scheme provided by default with the function PaletteWithoutGrey}

\item{usegradient}{whether the barplot should be draw in color gradient format}

\item{colorgradient}{the color scheme of gradient applied, default value c("#4575B4","ivory", "#D73027")}

\item{linegraph}{whether to plot the graph in a line graph format, default set to FALSE}

\item{log2scale}{whether the yscales should be in log2 scale, default set to TRUE}

\item{ratio}{aspect ratio of the plot, default set to 0.6}

\item{ret_plot}{Logical to tell if you want to return the bar plot from the protein selected}

\item{withpopup}{A logical, only use in shiny context. If TRUE, will call a modal dialog in a shiny app;
else, it will ask a question directly in the R console (readline function)}

\item{continue}{A logical to tell if you wanna continue and get the bar plots after getting the similar profile.
This was thought when withpopup is TRUE. If continue is FALSE will only stop after getting similar profiles.}

\item{modvar}{A character used when withpopup is TRUE, so in shiny context. Y or YES and the function goes on;
N or NO and the function stop (doesn't save and return the similar protein profiles)}

\item{got_it}{A logical to tell if you already have your data filtered. If TRUE, will directly starting to get the bar plot.}

\item{save_prlist}{A logical to tell if you want to save or not the protein list with a similar profile}

\item{save_pdf}{A logical to tell if you want to save plots in a pdf file}

\item{toplabel}{textual label at the top part of the page}

\item{leftlabel}{textual label at the left side of the page}

\item{bottomlabel}{textual label at the bottom part of the page}

\item{pdfname}{textual label of the pdf file}

\item{pdfheight}{a number indicate the height of pdf file, default value 12}

\item{pdfwidth}{a number indicate the width of pdf file, default value 12}
}
\value{
The ms 2D barplot
}
\description{
Function to generate 2D bar plot and pdf file with multipanel bar plots for 2D-CETSA data of
proteins which have similar profile from a selected protein.
It is based on the function imprints_barplotting from the IMPRINTS.CETSA package.
}
\seealso{
\code{\link{imprints_barplotting_app}} , \code{\link{imprints_corr_to_ref}}
}
