% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprints_categorize_peptides.R
\name{imprints_categorize_peptides}
\alias{imprints_categorize_peptides}
\title{imprints_categorize_peptides}
\usage{
imprints_categorize_peptides(
  data,
  data_cleaved,
  control,
  save_xlsx = TRUE,
  xlsxname = "RESP_summary_categorized"
)
}
\arguments{
\item{data}{The normalized peptides data set, i.e. the outpout from \code{imprints_normalize_peptides}.}

\item{data_cleaved}{The cleavage hits data set, i.e. the outpout from \code{imprints_cleaved_peptides}.}

\item{control}{The control treatment from your dataset.}

\item{save_xlsx}{Logical to tell if you want to save the categorized hits in an xlsx file.
Default to TRUE.}

\item{xlsxname}{The name of your saved file.}
}
\value{
A data.frame containing the categorized hits. It will add the two columns 'category' and 'details'
to the data.frame returned by the function \code{imprints_cleaved_peptides}.
}
\description{
Function to categorize proteins found as hits by the function \code{imprints_cleaved_peptides}
}
\details{
This function will categorize the hits found by the function \code{imprints_cleaved_peptides} in 4
main categories: RESP for REgional Stabilization after Proteolysis, SP for Single Peptide,
MP for multiple peptides and Unconfident for Unconfident. When a small 'm' is added to SP or MP it stands
for modified.
}
