% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMPRINTS_complex_mapping_sh.R
\name{IMPRINTS_complex_mapping_sh}
\alias{IMPRINTS_complex_mapping_sh}
\title{IMPRINTS_complex_mapping_sh}
\usage{
IMPRINTS_complex_mapping_sh(
  data,
  categorytable = NULL,
  set = NULL,
  treatment = NULL,
  complexdatabase = "Corum",
  organism = "Human",
  complexID = NULL,
  targetcategory = c("NC", "CN", "CC"),
  minsubunitsIdentified = 3,
  cutisoformbycount = TRUE
)
}
\arguments{
\item{data}{dataset after calculating the relative protein abundance differences and
deriving the average of the profile, ie IMPRINTS_average_sh,
make sure the columns with readings are named in the format like "37C_Treatment".}

\item{categorytable}{Dataset containing the category of each proteins from data, ie hitlist.
The column that contains the conditions must be named treatment.}

\item{set}{A single character to indicate the sample name if any}

\item{treatment}{A single character to indicate the sample name}

\item{complexdatabase}{By default is core Corum database, 03.09.2018 Corum 3.0 current release}

\item{organism}{Choose from Human/Mouse/Rat, default is Human}

\item{complexID}{A vector of complexID to retrieve from the data, if specified}

\item{targetcategory}{A character vector to select the category you want. Default value is c("NC","CN","CC")}

\item{minsubunitsIdentified}{The minimal number of subunits should be identified, default value is 3}

\item{cutisoformbycount}{When there are several isoforms mapped in a complex,
by default to filter away the isoforms with comparatively fewer count number}
}
\value{
A dataframe with the assigned protein complex
}
\description{
Function to generate 2D bar plot and pdf file with multipanel bar plots for 2D-CETSA data of
proteins which have similar profile from a selected protein.
It is totally based on the function ms_2D_barplotting from the mineCETSA package.
}
