% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprints_IS.R
\name{imprints_IS}
\alias{imprints_IS}
\title{imprints_IS}
\usage{
imprints_IS(
  data,
  data_diff = NULL,
  ctrl,
  valid_val = NULL,
  IS_cutoff = 1.5,
  fixed_score_cutoff = FALSE,
  FDR = 0.01,
  pv_method = c("all", "top2"),
  adj_pv_method = "BH",
  comb_pv_method = c("fisher", "george", "edgington"),
  curvature = 0.05,
  FDR_category = 0.1,
  format_category = c("9", "4"),
  folder_name = "Hits_analysis",
  peptide_count_col = "sumUniPeps",
  species = "Homo Sapiens"
)
}
\arguments{
\item{data}{The input data set from which categorization is performed on and hitlist is produced from.}

\item{data_diff}{The output from imprints_caldiff; can be NULL and if so will compute it; can also be a path to the data.}

\item{ctrl}{The name of the control.}

\item{valid_val}{The percentage of non-missing values you want per treatment. If less, score will be set to NA,
i.e. the protein will not be a hit}

\item{IS_cutoff}{The I-score cutoff. Default is 1.5.}

\item{fixed_score_cutoff}{Logical to tell if you want to use a fixed cutoff for the I-score.
If TRUE, the value IS_cutoff will directly be used as the cutoff and for all treatments. If FALSE,
the I-score cutoff will be calculated as the value selected for IS_cutoff plus the median of the
I-scores of the proteins which have a p-value lower than the median of all p-values for a given treatment.
Default is FALSE.}

\item{FDR}{The FDR used for the BH corrected combined p-value}

\item{pv_method}{The method to compute the p-values. If top2, then only the p-values from the two greatest fold-changes
will be computed and combined; if all, will take all fold-changes.
Default is all.}

\item{adj_pv_method}{see \code{\link{p.adjust}}; default is BH}

\item{comb_pv_method}{The method used to combine p-values. Either george, fisher or edgington.
Default is fisher; see Details.}

\item{curvature}{The curvature used for the curve on the volcano plot}

\item{FDR_category}{The FDR used for the BH corrected  p-value at 37°C used in order to categorize the hits}

\item{format_category}{Choose between 9 or 4, indicating how many categories to segregate the hits; default value is 9.
If 9  is selected, the 9 categories will be: NN, CN+, CN-, NC+, NC-, CC++, CC+-, CC-+, CC–.
If 4, then it will be: NN, CN, NC, CC.
The sign of + or - after N or C is determined by the sign of the fold-change at 37°C and the signof the I-score, respectively.}

\item{folder_name}{The name of the folder in which you want to save the results.}

\item{peptide_count_col}{The name of the column that contain the unique peptide count.
If it is sumUniPeps, don't bother with this parameter.}

\item{species}{The species on which you did the experiment (not necessary if already present in your datas).
Default is 'Homo Sapiens'.}
}
\value{
A dataframe which contains the hits.
}
\description{
Function to categorize protein according to their Expression/Stability change
based on their I-score and their combined p-value on their two biggest fold changes.
}
\details{
George's method correspond to the sum of the logit of the p-values, Fisher's to the sum of the
log of the p-values and Edgington's is the sum of the p-values.
Edgington's method is the most stringent and is particularly sensitive with higher p-values
whereas Fisher's method is the less stringent as it is mostly sensitive to low p-values.
George's method is a compromise between the two methods.
For more details read \link{https://doi.org/10.48550/arXiv.1707.06897}.
}
