% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicing_form_mapping.R
\name{imprints_isoform_peptides}
\alias{imprints_isoform_peptides}
\title{imprints_isoform_peptides}
\usage{
imprints_isoform_peptides(
  data,
  data_cleaved,
  control,
  fasta,
  minimum_align = 5,
  save_xlsx = TRUE,
  xlsxname = "RESP_isoform_mapping"
)
}
\arguments{
\item{data}{The normalized peptides data set, i.e. the outpout from \code{imprints_normalize_peptides}.}

\item{data_cleaved}{The cleavage hits data set, i.e. the outpout from \code{imprints_cleaved_peptides}.}

\item{control}{The control treatment from your dataset.}

\item{fasta}{The path to the FASTA file you used for the search}

\item{minimum_align}{Numeric to tell the minimum aligned sequence length. Default to 5.
It is advised to put the same as the minimum peptide sequence length you selected in
the proteomics search software you used.}

\item{save_xlsx}{Logical to tell if you want to save the categorized hits in an xlsx file.
Default to TRUE.}

\item{xlsxname}{The name of your saved file.}
}
\value{
A dataframe containing the potential splicing forms which could be more
        expressed instead of the protein being cleaved
}
\description{
Function to check if the hits found by \code{\link{imprints_cleaved_peptides}} could be due to alternate
splicing forms being more expressed and not due to protein cleavage.
}
\details{
When a hit is returned by \code{\link{imprints_cleaved_peptides}}, it means that this protein has a peptide position
where the IMPRINTS profiles of the two obtained parts are significantly different. This difference can be caused
by protein modification and mainly proteolysis; but if a splicing form of protein is more expressed than its canonical
form, a significant difference in the profiles can also occur. The aim here is to refilter the hit list and give
the possible splicing forms which could be more expressed based on the output of \code{\link{imprints_cleaved_peptides}}
and the sequence alignments of the isoform sequence and its corresponding canonical form.
}
\seealso{
\code{\link{imprints_cleaved_peptides}}
}
