% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprints_join_peptides.R
\name{imprints_join_peptides}
\alias{imprints_join_peptides}
\title{imprints_join_peptides}
\usage{
imprints_join_peptides(data)
}
\arguments{
\item{data}{A list of peptides data to join, typically after \code{imprints_sequence_peptides}}
}
\value{
Your filtered dataset
}
\description{
Function to join peptides dataset obtained after \code{imprints_sequence_peptides}.
}
