% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprints_network.R
\name{imprints_network}
\alias{imprints_network}
\title{imprints_network}
\usage{
imprints_network(
  data,
  hits = NULL,
  treatment = NULL,
  GOterm = NULL,
  required_score = 400,
  node_wolink = TRUE,
  witherrorbar = TRUE,
  FC_border = TRUE,
  colorbar = NULL,
  colorFC = c("#0041FF", "#FFFFFF", "#FF0000"),
  species = c("human", "mouse", "rat"),
  physics_type = c("forceAtlas2Based", "barnesHut", "repulsion", "hierarchicalRepulsion"),
  physics_enabled = TRUE,
  ...
)
}
\arguments{
\item{data}{Dataset after imprints_caldiff}

\item{hits}{The Uniprots ids from which you want to plot the network.
If NULL, it will take all the proteins from the data.}

\item{treatment}{The treatment from which you want to plot to plot the barplots.
If NULL, it will plot all the treatments present in the data.}

\item{GOterm}{Either a data frame containing the columns 'id', 'GOterm' and eventually 'color';
a character corresponding to a data base or a NULL object.
If you choose a data base, it will perform an enrichment analysis and assign a GO term
with the lowest FDR to each protein.
The data bases now supported are: COMPARTMENTS, Process, Component, Function, TISSUES,
Keyword, KEGG, SMART, PMID, RCTM, WikiPathways and NetworkNeighborAL}

\item{required_score}{A numeric between 0 and 1000 to set the required score
for the interectation in the STRING network.}

\item{node_wolink}{Logical to tell if you want to print node without links in the network.
Default set to TRUE}

\item{witherrorbar}{Logical to tell if you want to print the error bar or not on the bar plots.}

\item{FC_border}{Logical to tell if you want to color the nodes borders according to the
maximum Fold-Change from each protein.}

\item{colorbar}{A vector of colors corresponding to each treatment.}

\item{colorFC}{If FC_border is set to TRUE, a vector of three color corresponding
to the min, mid and max value from the maximum fold change.}

\item{species}{The species; either human, mouse or rat.}

\item{physics_type}{A string corresponding to the solver of the network; default is forceAtlas2Based.
Possible options: 'barnesHut', 'repulsion', 'hierarchicalRepulsion', 'forceAtlas2Based'.}

\item{physics_enabled}{Logical to toggle the physics system on or off.}

\item{...}{Options for network visualization. Options available are:
borderWidth, imagePadding, node_size, edge_length, edge_color
font_background, font_color, legend_width, file_type}
}
\value{
visNetwork object
}
\description{
Function to plot an interactive STRING network with the IMPRINTS barplots in each nodes
}
