% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprints_normalize_peptides.R
\name{imprints_normalize_peptides}
\alias{imprints_normalize_peptides}
\title{imprints_normalize_peptides}
\usage{
imprints_normalize_peptides(peptides_data, dataset_name = "imprints")
}
\arguments{
\item{peptides_data}{The data frame output from \code{imprints_read_peptides}. Can also be the path to that file.}

\item{dataset_name}{The name of your dataset}
}
\value{
The normalized peptides data frame, non log2 transformed.
}
\description{
Function to normalize your peptides IMPINTS-CETSA dataset.
}
\details{
The function uses the function \code{imprints_normalization} from \code{IMPRINTS.CETSA} package.
         The function apply the VSN model to each temperature set and then call the function \code{removeBatchEffect}
         from \code{limma} package.
}
