% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprints_plotting_isoform_peptides.R
\name{imprints_plotting_isoform_peptides}
\alias{imprints_plotting_isoform_peptides}
\title{imprints_plotting_isoform_peptides}
\usage{
imprints_plotting_isoform_peptides(
  data,
  data_isoform,
  control,
  treatment,
  min_ValidValue = 0.4,
  ret_plot = TRUE,
  save_pdf = FALSE,
  layout = c(1, 2),
  toplabel = "",
  leftlabel = "",
  bottomlabel = "",
  pdfname = "isoform_alignement_plots",
  pdfheight = 9,
  pdfwidth = 18
)
}
\arguments{
\item{data}{The normalized peptides data set, i.e. the outpout from \code{imprints_normalize_peptides}.}

\item{data_isoform}{The cleavage hits data aligned with the potential isoforms, i.e. the outpout from \code{imprints_isoform_peptides}.}

\item{control}{The control treatment from your dataset.}

\item{treatment}{The treatment from which you want to see the peptides fold-changes. Can only be one.}

\item{min_ValidValue}{The minimum proportion of non-missing values per peptides.
Default is 0.4; so if 6 temperatures need at least 3 non missing values.}

\item{ret_plot}{Logical to tell if you want to return the last plot}

\item{save_pdf}{A logical to tell if you want to save plots in a pdf file}

\item{layout}{A vector indicating the panel layout for multi-panel plots per page,
default value is c(2,2), use when save_pdf = TRUE}

\item{toplabel}{textual label at the top part of the page}

\item{leftlabel}{textual label at the left side of the page}

\item{bottomlabel}{textual label at the bottom part of the page}

\item{pdfname}{Character to label the pdf file}

\item{pdfheight}{Numeric indicating the height of pdf file, default value 12}

\item{pdfwidth}{Numeric indicating the width of pdf file, default value 12}
}
\value{
The plots of the isoform sequence alignment
}
\description{
Function to visualize the isoform sequence alignment with its canonical form obtained from
the function \code{\link{imprints_isoform_peptides}} to compare with the cleavage site obtained
by \code{\link{imprints_cleaved_peptides}}.
}
\details{
The plot will show you how each peptides aligns with the canonical sequence. Each peptide will be clustered
in two groups, the one with greatest log2 fold-change and the ones with the lowest.
This has been thought for a final control of the results from \code{\link{imprints_isoform_peptides}} and
\code{\link{imprints_cleaved_peptides}} in order to see if a protein having a significant RESP effect could
be due to the expression of an alternative splicing form.
}
\seealso{
\code{\link{imprints_isoform_peptides}}
}
