% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprints_sequence_peptides.R
\name{imprints_sequence_peptides}
\alias{imprints_sequence_peptides}
\title{imprints_sequence_peptides}
\usage{
imprints_sequence_peptides(
  data,
  proteins = NULL,
  sequence = NULL,
  control = "Vehicle",
  barplot = FALSE,
  dataset_name = "imprints"
)
}
\arguments{
\item{data}{The input data set, the peptide data file filtered (non TMT modification removed) and non log2 transformed.}

\item{proteins}{The proteins you want to select. If NULL, select all.}

\item{sequence}{The peptide position you want to select. It needs to be a list that contains the same
number of element than the number of proteins or only one element. An element can be a string
like 191-199 or a vector that contains several strings like \code{c("52-71", "207-222")}.
Anyway, it needs to be in the format number-number.
If it's NULL, it will only select according the proteins so you can plot all the peptides from those.}

\item{control}{The condition corresponding to the control in your dataset.}

\item{barplot}{Logical to tell if you want to plot the barplots from the obtained fold-changes. Default is FALSE}

\item{dataset_name}{The name of your dataset to save your file.}
}
\value{
The caldiff output from the peptide position from the protein you selected; save it and also save the barplots.
}
\description{
Function to select proteins and sequence from peptide data, sum up the peptide that are not selected
(before and after the sequence selected) and then plot the fold change on a pdf.
}
