% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_in_zone.R
\name{is_in_zone}
\alias{is_in_zone}
\title{is_in_zone}
\usage{
is_in_zone(border, target)
}
\arguments{
\item{border}{A dataframe that contains the coordinates of the points that delimit the zone (a square for example)}

\item{target}{The coordinates of the points you want to check if it is in the zone}
}
\value{
A logical to tell if the point is in the zone
}
\description{
Function to tell if a point is in a delimited zone or not.
}
\examples{
square <- data.frame(x = c(1,1,2,2), y = c(1,2,2,1))
is_in_zone(square, c(1.5,1.5))

}
