% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_cetsa.R
\name{join_cetsa}
\alias{join_cetsa}
\title{join_cetsa}
\usage{
join_cetsa(list_data, new_names = c("1h", "6h"))
}
\arguments{
\item{list_data}{A list of data frames. Usually the data frames are the output from the imprints_caldiff function.}

\item{new_names}{The new character element you want to add at the end of the columns name.}
}
\value{
The joined data frame (by 'id' and 'description')
}
\description{
Function to join a list of data frame from CETSA experiment and renamed the column.
(Allow to avoid the adding of ".x" or ".y" at the end of the column names)
}
