% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{load_data}
\usage{
load_data()
}
\value{
The list drug_data
}
\description{
Function to load the data saved in the drug_data folder. Will import in the drug_data format.
}
