% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mineCETSAapp.R
\docType{package}
\name{mineCETSAapp}
\alias{mineCETSAapp}
\title{mineCETSAapp Package}
\description{
Contains a shiny app for mineCETSA R package, in a user friendly way.
Also contains several functions on which the app depends.
}
\author{
Marc-Antoine GERAULT \email{marco.gerault@gmail.com}
}
