% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NN_chemarrest}
\alias{NN_chemarrest}
\title{The non hits from the chemarrest dataset}
\format{
## `NN_chemarrest`
A data frame with 4706 rows and 4 columns:
\describe{
  \item{id}{Uniprot protein ID}
  \item{description}{Protein description}
  \item{treatment}{The treatment in which it was identified as a hit}
  \item{category}{Category of the non hit, i.e. NN}
}
}
\source{
This data were obtained from 7.	Dai, L., T. Zhao, X. Bisteau et al. (2018)
        Modulation of Protein-Interaction States through the Cell Cycle.
        Cell, 173, 1481-1494.e13.
}
\usage{
NN_chemarrest
}
\description{
The non hits derived from the hitlist of the chemarrest dataset
}
\keyword{datasets}
