% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NN_elutriation}
\alias{NN_elutriation}
\title{The non hits from the elutriation dataset}
\format{
## `NN_elutriation`
A data frame with 4567 rows and 4 columns:
\describe{
  \item{id}{Uniprot protein ID}
  \item{description}{Protein description}
  \item{treatment}{The treatment in which it was identified as a hit}
  \item{category}{Category of the non hit, i.e. NN}
}
}
\source{
This data were obtained from 7.	Dai, L., T. Zhao, X. Bisteau et al. (2018)
        Modulation of Protein-Interaction States through the Cell Cycle.
        Cell, 173, 1481-1494.e13.
}
\usage{
NN_elutriation
}
\description{
The non hits derived from the hitlist of the elutriation dataset
}
\keyword{datasets}
