% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_in_orga.R
\name{point_in_orga}
\alias{point_in_orga}
\title{point_in_orga}
\usage{
point_in_orga(Organelle = "Golgi")
}
\arguments{
\item{Organelle}{One of the valid organelle; To see the valid organelles type
\code{unique(IMPRINTS.CETSA.app::orgatlas_match$location.corres)}}
}
\value{
A double containing coordinates
}
\description{
Function to generate random coordinate according to a subcellular location from img2.
}
\seealso{
\code{\link{is_in_zone}}
}
