% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pr_atlas}
\alias{pr_atlas}
\title{Human protein cellular localization from Protein Atlas database}
\format{
## `pr_atlas`
A data frame with 17 377 rows and 6 columns:
\describe{
  \item{Gene}{Gene}
  \item{Uniprot}{Uniprot ID}
  \item{Reliability..IH.}{Reliability..IH.}
  \item{Subcellular.location}{Subcellular.location}
  \item{Subcellular.main.location}{Subcellular.main.location}
  \item{Subcellular.additional.location}{Subcellular.additional.location}
}
}
\source{
https://www.proteinatlas.org/api/search_download.php?search=Human&format=json&columns=g,up,relih,scl,scml,scal&compress=no
  Last downloaded on the 21th of December 2023
}
\usage{
pr_atlas
}
\description{
Human protein cellular localization from Protein Atlas database
}
\keyword{datasets}
