% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remData.R
\name{remData}
\alias{remData}
\title{remData}
\usage{
remData(data, old_drug)
}
\arguments{
\item{data}{A list of data (the data saved in the folder drug_data)}

\item{old_drug}{The name of the drug you want to remove (has to be in the drug_data folder)}
}
\description{
Function to remove a drug from the data in the folder drug_data.
}
