% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCETSAapp.R
\name{runCETSAapp}
\alias{runCETSAapp}
\title{Run the app to use IMPRINTS.CETSA.app package}
\usage{
runCETSAapp()
}
\description{
\code{runCETSAapp} works exactly the same as runExample from \code{\link{shiny}} package.
}
