% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveData.R
\name{saveData}
\alias{saveData}
\title{saveData}
\usage{
saveData(data, new_add = NULL, new_drug = NULL)
}
\arguments{
\item{data}{A list of data that you want to save}

\item{new_add}{A list of new data you want to add to the existing xlsx files}

\item{new_drug}{A character vector which is the name of the new drug you're adding}
}
\description{
Function for saving new data in the folder drug_data.
If in the drug_data format, you can add new data from another drug.
}
