% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR_CetsaHit.R
\name{SR_CetsaHit}
\alias{SR_CetsaHit}
\title{SR_CetsaHit}
\usage{
SR_CetsaHit(
  data,
  data_diff = NULL,
  ctrl,
  valid_val = NULL,
  SR_cutoff = 1,
  pval_cutoff = 0.05,
  padj_method = "BH",
  folder_name = "Hits_analysis"
)
}
\arguments{
\item{data}{The input data set from which categorization is performed on and hitlist is produced from.}

\item{data_diff}{The output from ms_2D_caldiff; can be NULL and if so will compute it; can also be a path to the data.}

\item{ctrl}{The name of the control.}

\item{valid_val}{The percentage of valid values you want per condition. If less, score will be set to NA, 
i.e. the protein will not be a hit}

\item{SR_cutoff}{The stability rate cutoff}

\item{pval_cutoff}{The combined p-value cutoff}

\item{padj_method}{The p-value adjust method, default is BH.}

\item{folder_name}{The name of the folder in which you want to save the results.}
}
\value{
A dataframe which contains the hit.
}
\description{
Function to categorize protein according to their Expression/Stability change
based on their stability rate and their combined p-value on their two biggest fold changes.
}
\details{
This function is still on test. Your data should contain the columns named 
         Protein.Group, Genes and peptides_counts_all.
}
