% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cetsa_gsea.R
\name{cetsa_gsea}
\alias{cetsa_gsea}
\title{cetsa_gsea}
\usage{
cetsa_gsea(
  hits,
  gene_column = "Genes",
  score_column = "SR",
  condition_column = NULL,
  condition = NULL,
  species = c("human", "mouse"),
  pos_enrichment = TRUE,
  pval_cutoff = 0.01,
  database = c("WikiPathway", "KEGG", "GO")
)
}
\arguments{
\item{hits}{A data.frame containing the genes id, a score value and preferably a condition column but not necessary.}

\item{gene_column}{The name of the coulumn that contains the genes. Default is 'Genes'.}

\item{score_column}{The name of the coulumn that contains the score values. Default is 'SR'.}

\item{condition_column}{The name of the column that contains the conditions. Default is NULL.}

\item{condition}{The name of the condition you ant to keep. Default is NULL.}

\item{species}{Specify the species. Currently, only 'human' and 'mouse' are available.}

\item{pos_enrichment}{Logical to tell if you want to only look at positive enrichment score.
If FALSE, show only negative.}

\item{pval_cutoff}{The p-value cutoff for the enrichment analysis.}

\item{database}{Specify the database. Currently, WikiPathway, KEGG and GO ar available.}
}
\value{
A list that contains the results and the plot.
}
\description{
Function to run enrichment analysis and plot a GSEA plot.
}
\seealso{
\code{\link{clusterProfiler}}
}
