% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_conceptNet.R
\name{gene_conceptNet}
\alias{gene_conceptNet}
\title{gene_conceptNet}
\usage{
gene_conceptNet(
  hits,
  gene_column = "Genes",
  score_column = "SR",
  condition_column = NULL,
  condition = NULL,
  species = c("human", "mouse"),
  pval_cutoff = 0.01,
  database = c("WikiPathway", "KEGG", "GO")
)
}
\arguments{
\item{hits}{A data.frame containing the genes id, a score value and preferably a condition column but not necessary.}

\item{gene_column}{The name of the coulumn that contains the genes. Default is 'Genes'.}

\item{score_column}{The name of the coulumn that contains the score values. Default is 'SR'.}

\item{condition_column}{The name of the column that contains the conditions. Default is NULL.}

\item{condition}{The name of the condition you ant to keep. Default is NULL.}

\item{species}{Specify the species. Currently, only 'human' and 'mouse' are available.}

\item{pval_cutoff}{The p-value cutoff for the gene concept network.}

\item{database}{Specify the database. Currently, WikiPathway, KEGG and GO ar available.}
}
\value{
The gene concept network plot.
}
\description{
Function to run enrichment analysis on your hits and return
a network plot.
}
\seealso{
\code{\link{clusterProfiler}}
}
