% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitlist.R
\name{hitlist}
\alias{hitlist}
\title{Hitlist}
\usage{
hitlist(
  inputdata,
  meancutoff = 0.25,
  boundedness = 4,
  qualitycutoff = 0.15,
  meandev = 0.15,
  expstr = c(),
  use_prompt = TRUE,
  exported = FALSE
)
}
\arguments{
\item{inputdata}{The input data set from which categorization is performed on and hitlist is produced from.}

\item{meancutoff}{The threshold value for the mean of a triplicate(First filter of hits).}

\item{boundedness}{Number of SD/SEM, the mean needs to be within.}

\item{qualitycutoff}{The threshold value for the SD/SEM to separate well-measured and illmeasured measurements.}

\item{meandev}{The deviation of individual mean from the group mean in CN categorization.}

\item{expstr}{List of strings with treatments/drugs and control group}

\item{use_prompt}{Logical to tell if you want to use the prompt function (added for the shiny app)}

\item{exported}{Logical to tell if you want to export the results. Use it when use_prompt is FALSE.}
}
\value{
Dataframe of the export.
}
\description{
Function to categorize protein according to their Expression/Stability change
based on two treatment conditions and characteristic conditions for fold change.
}
