% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprints_barplotting_categorize_peptides.R
\name{imprints_barplotting_categorize_peptides}
\alias{imprints_barplotting_categorize_peptides}
\title{imprints_barplotting_categorize_peptides}
\usage{
imprints_barplotting_categorize_peptides(
  data,
  data_cleaved,
  treatment,
  control,
  format = c("RESP_peptide", "peptide_one"),
  color = "red",
  pdfname = "categorized_RESP_barplots"
)
}
\arguments{
\item{data}{The normalized peptides data set, i.e. the outpout from \code{imprints_normalize_peptides}.}

\item{data_cleaved}{The categorized cleavage hits data set, i.e. the outpout from \code{imprints_categorize_peptides}.}

\item{treatment}{The treatment from which you want to see the plots. Can only be one.}

\item{control}{The control treatment from your dataset.}

\item{format}{Format of the plot; either \code{RESP_peptide} which will plot the RESP plot and then all peptides
separately from the protein or \code{peptide_one} which will plot all peptides in one plot for each protein.
Default is \code{RESP_peptide}.}

\item{color}{The color of the bar plots.}

\item{pdfname}{textual label of the pdf file}
}
\value{
NULL. Will save a pdf file with ordered and catgorized plots
}
\description{
Function to plot the categorized proteins found as hits by the function \code{imprints_cleaved_peptides}
}
\seealso{
\code{\link{imprints_barplotting_peptides}}

\code{\link{imprints_categorize_peptides}}
}
