% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprints_cleaved_peptides.R
\name{imprints_cleaved_peptides}
\alias{imprints_cleaved_peptides}
\title{imprints_cleaved_peptides}
\usage{
imprints_cleaved_peptides(
  data,
  data_diff = NULL,
  control = "Vehicle",
  min_ValidValue = 0.4,
  FDR = 0.01,
  diff_cutoff = 0.25,
  curvature = 0.1
)
}
\arguments{
\item{data}{The normalized peptides data set, i.e. the outpout from \code{imprints_normalize_peptides}.}

\item{data_diff}{The log2 fold-changes peptides data set, i.e. the outpout from \code{imprints_sequence_peptides}.
If NULL, it will be computed.}

\item{control}{The control treatment from your dataset.}

\item{min_ValidValue}{The minimum proportion of valid values per peptides.
Default is 0.4; so if 6 temperatures need at least 3 non missing values.}

\item{FDR}{The FDR used to obtained the final p-value cutoff and maximum fold-change cutoff. Default is 1%}

\item{diff_cutoff}{The maximum fold-change cutoff. Default is 0.25}

\item{curvature}{The curvature used for the curve on the volcano plot}
}
\value{
The potential cleaved sites from the proteins considered as cleaved.
}
\description{
Function to find proteins that are potentially cleaved and their cleaved sites.
For more information, see Details section.
}
\details{
The idea here is to compute the sum from all fold change for each peptide and then compute the cumulative sum of
these summed fold change of every peptides for each protein.
If the protein is not cleaved, this cumulative sum should be constantly increasing/decreasing, i.e linear.
If the protein is cleaved, then this cumulative sum will either be convex or concave.
If convex, the last peptides are more abundant; if concave the first peptides are more abundant.
In the end a volcano plot will be plotted where the p-value corresponds to the combined p-value from
the six temperatures comparison between peptides located towards the N-terminal and the ones located
towards the C-terminal position.

You can then use the function \code{imprints_sequence_peptides} to plot the peptides before and after the potential cleaved site found.
}
