% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprints_cleaved_peptides.R
\name{imprints_cleaved_peptides}
\alias{imprints_cleaved_peptides}
\title{imprints_cleaved_peptides}
\usage{
imprints_cleaved_peptides(data, R2 = 0.9, control = NULL)
}
\arguments{
\item{data}{The input data set, i.e. the outpout from \code{imprints_sequence_peptides}.}

\item{R2}{The R-squared cutoff. It characterize the linearity of the cumulative sum of the fold changes
from every peptide from each protein. The higher it is, the less stringent you are on the decision
of taking a protein as cleaved. Default is 0.9.}

\item{control}{The control condition from your dataset. If the condition is found in your data, then it is removed from it.}
}
\value{
The potential cleaved sites from the proteins considered as cleaved.
}
\description{
Function to find proteins that are potentially cleaved and their cleaved sites.
For more informations, see Details section.
}
\details{
Let's say you found some hits in your dataset and you want to check if some of those
are cleaved. You first compute the fold changes for every peptides from each hit and then
search for potential cleaved sites.
The idea here is to compute the sum from all fold change for each peptide and then compute the cumulative sum of
these summed fold change of every peptides for each protein.
If the protein not cleaved, this cumulative sum should be constantly increasing/decreasing, i.e linear.
If the protein is cleaved, then this cumulative sum will either be convex or concave.
If convex, the last peptides are more abundant; if concave the first peptides are more abundant.

The potential cleaved sites found, can only be peptides that are in the dataset.
Also, it's better to double check by plotting the peptides before and after the potential cleaved site found
by using \code{imprints_sequence_peptides}.
}
