% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprints_join_peptides.R
\name{imprints_join_peptides}
\alias{imprints_join_peptides}
\title{imprints_join_peptides}
\usage{
imprints_join_peptides(data)
}
\arguments{
\item{data}{A list of peptides data to join, typically after \code{imprints_sequence_peptides}}
}
\value{
Your filtered dataset
}
\description{
Function to join peptides dataset obtained after \code{imprints_sequence_peptides}.
Although, this function will only join on the protein position and will try to join close
position together; i.e. it will not take into account the amino-acid sequence.
Indeed, this function was written to join peptides dataset that would contain summed peptides
fold-changes before and after a specific position.
}
