% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprints_join_peptides.R
\name{imprints_join_peptides}
\alias{imprints_join_peptides}
\title{imprints_join_peptides}
\usage{
imprints_join_peptides(data, mode = c("partial", "exact"))
}
\arguments{
\item{data}{A list of peptides data to join, typically after \code{imprints_sequence_peptides}}

\item{mode}{Either 'exact' or 'partial'. If 'exact', the datasets will be joined according their
peptide sequence, i.e. to be joined they need to have the same annotated sequence, position and modification.
If 'partial', peptides will only be joined on the protein position and will try to join close
position together; i.e. it will not take into account the amino-acid sequence. It is mostly useful
for joining dataset after summing different peptides parts together using the function \code{imprints_sequence_peptides}.
Default is 'partial'.}
}
\value{
Your joined dataset
}
\description{
Function to join peptides dataset obtained after \code{imprints_sequence_peptides}.
}
