% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprints_read_peptides.R
\name{imprints_read_peptides}
\alias{imprints_read_peptides}
\title{imprints_read_peptides}
\usage{
imprints_read_peptides(
  peptides_files,
  treatment,
  temperatures,
  proteins,
  modification_tokeep = c("TMT"),
  dataset_name = "Venetoclax6h"
)
}
\arguments{
\item{peptides_files}{The path to the peptides data. Can be one or more files.}

\item{treatment}{A character vector that contains the treatment applied for each channel like 'B1_Vehicle' for example.}

\item{temperatures}{A character vector that contains the temperatures used for each peptides file like '37C' for example.}

\item{proteins}{Either a data frame or a file that has the 2 columns 'id' and 'description' corresponding
to the proteins you want to keep (usually proteins after \code{imprints_rearrange} or \code{imprints_caldiff})}

\item{modification_tokeep}{A character vector that contains the peptides modifications you want to keep. Default is 'TMT'}

\item{dataset_name}{The name of your dataset}
}
\value{
The peptides data frame
}
\description{
Function to read your peptides files and concatenate them in one data frame.
}
