% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprints_read_peptides.R
\name{imprints_read_peptides}
\alias{imprints_read_peptides}
\title{imprints_read_peptides}
\usage{
imprints_read_peptides(
  peptides_files,
  treatment,
  temperatures,
  prefixcontaminant = "",
  averagecount = TRUE,
  countthreshold = 2,
  proteins = NULL,
  modification_torm = NULL,
  dataset_name = "imprints"
)
}
\arguments{
\item{peptides_files}{The path to the PeptidesGroups file from PD. Can be one or more files.}

\item{treatment}{A character vector that contains the treatment applied for each channel like 'B1_Vehicle' for example.}

\item{temperatures}{A character vector that contains the temperatures used for each peptides file like '37C' for example.}

\item{prefixcontaminant}{Character corresponding to the prefix used to identify contaminants.}

\item{averagecount}{Whether to take the median of the abundance count numbers across the measured temperature
range and then use this value for filtering. Default set to TRUE.
Otherwise, filter the  peptides according to the associated count numbers at each temperature.}

\item{countthreshold}{The minimal threshold number of associated abundance count of peptides, default is 2.}

\item{proteins}{Either a data frame or a file that has the 2 columns 'id' and 'description' corresponding
to the proteins you want to keep (usually proteins after \code{imprints_rearrange} or \code{imprints_caldiff}).
If not specified, no filtering is applied and you must have the column 'Master Protein Descriptions' in your PeptideGroup file(s).}

\item{modification_torm}{A character vector that contains the peptides modifications you want to remove. Default is NULL}

\item{dataset_name}{The name of your dataset}
}
\value{
The peptides data frame
}
\description{
Function to read your peptides files and concatenate them in one data frame.
}
