% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprints_remove_peptides.R
\name{imprints_remove_peptides}
\alias{imprints_remove_peptides}
\title{imprints_remove_peptides}
\usage{
imprints_remove_peptides(data, proteins = NULL, sequence)
}
\arguments{
\item{data}{A peptides dataset, typically after \code{imprints_sequence_peptides}.
It needs to contains the columns 'Master Protein Accessions' and 'Positions in Master Proteins'.}

\item{proteins}{The proteins from which you want to remove the peptides.
If NULL and you put one sequence, it will remove all these peptides for all proteins from peptides data.
If NULL and you put several sequence, the number of sequence you put needs to match the number
of proteins you have in your peptides data.}

\item{sequence}{The peptide position you want to remove. If you put one sequence, it will remove it for all proteins
from your peptides data; otherwise it needs to match the number of proteins you put or have in your dataset.
The format needs to be a number followed by a dash and another number, like this : '208-221'.}
}
\value{
Your filtered dataset
}
\description{
Function to filter some peptides according there sequence. For example after using \code{imprints_sequence_peptides},
if you want to remove the peptides which corresponds to the cleaved sites.
}
