% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gsea.R
\name{run_gsea}
\alias{run_gsea}
\title{run_gsea}
\usage{
run_gsea(
  hits,
  gene_column = "Gene",
  score_column = "SR",
  treatment_column = NULL,
  treatment = NULL,
  species = c("human", "mouse"),
  pos_enrichment = TRUE,
  pval_cutoff = 0.01,
  database = c("WikiPathway", "KEGG", "GO")
)
}
\arguments{
\item{hits}{A data.frame containing the genes id, a score value and preferably a treatment column but not necessary.}

\item{gene_column}{The name of the coulumn that contains the genes. Default is 'Gene'.}

\item{score_column}{The name of the coulumn that contains the score values. Default is 'SR'.}

\item{treatment_column}{The name of the column that contains the treatments. Default is NULL.}

\item{treatment}{The name of the treatment you want to keep. Default is NULL.}

\item{species}{Specify the species. Currently, only 'human' and 'mouse' are available.}

\item{pos_enrichment}{Logical to tell if you want to only look at positive enrichment score.
If FALSE, show only negative.}

\item{pval_cutoff}{The p-value cutoff for the enrichment analysis.}

\item{database}{Specify the database. Currently, WikiPathway, KEGG and GO ar available.}
}
\value{
A list that contains the results and the plot.
}
\description{
Function to run enrichment analysis and plot a GSEA plot.
}
\seealso{
\code{\link{clusterProfiler}}
}
