% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SR_CetsaHit.R
\name{SR_CetsaHit}
\alias{SR_CetsaHit}
\title{SR_CetsaHit}
\usage{
SR_CetsaHit(
  data,
  data_diff = NULL,
  ctrl,
  valid_val = NULL,
  SR_cutoff = 1.5,
  FDR = 0.01,
  curvature = 0.1,
  FDR_category = 0.1,
  folder_name = "Hits_analysis",
  peptide_count_col = "peptides_counts_all",
  species = "Homo Sapiens"
)
}
\arguments{
\item{data}{The input data set from which categorization is performed on and hitlist is produced from.}

\item{data_diff}{The output from imprints_caldiff; can be NULL and if so will compute it; can also be a path to the data.}

\item{ctrl}{The name of the control.}

\item{valid_val}{The percentage of valid values you want per condition. If less, score will be set to NA,
i.e. the protein will not be a hit}

\item{SR_cutoff}{The stability rate cutoff}

\item{FDR}{The FDR used for the BH corrected combined p-value}

\item{curvature}{The curvature used for the curve on the volcano plot}

\item{FDR_category}{The FDR used for the BH corrected  p-value at 37°C used in order to categorize the hits}

\item{folder_name}{The name of the folder in which you want to save the results.}

\item{peptide_count_col}{The name of the column that contain the unique peptide count.
If it is sumUniPeps, don't bother with this parameter.}

\item{species}{The species on which you did the experiment (not necessary if already present in your datas).
Default is 'Homo Sapiens'.}
}
\value{
A dataframe which contains the hits.
}
\description{
Function to categorize protein according to their Expression/Stability change
based on their stability rate and their combined p-value on their two biggest fold changes.
}
\details{
This function is still on test. Your data should contain the columns named
         Protein.Group, Genes and peptides_counts_all.
}
